refactor(core-weave): extract extraction source block helpers
