+++ src/core/weave/weave.ts
@@ -23,7 +23,6 @@ import {
 } from "./support_history_policy.ts";
 import type { VersionPlan } from "./version_plan.ts";
 import type {
-  ExtractionSourceEvidenceModel,
   ResourcePageDefinitionWorkingArtifact,
   WeaveableKnopCandidate,
 } from "./candidates.ts";
@@ -71,9 +70,12 @@ import {
   toAbsoluteIri,
   toMeshRelativePath,
 } from "./rdf_helpers.ts";
-import { findSubjectBlockIndex, splitTurtleBlocks } from "./turtle_blocks.ts";
 import { classifyWeaveSlice } from "./slice_classification.ts";
 import { extractCurrentReferenceCatalogLinks } from "./reference_catalog_links.ts";
+import {
+  renderExactExtractionSourceBlock,
+  replaceExtractionSourceBlock,
+} from "./extraction_source_blocks.ts";
 import {
   buildCurrentOnlyReferenceCatalogWeavePages,
   buildFirstExtractedKnopWeavePages,
@@ -1572,115 +1574,6 @@ function resolveCurrentMeshInventoryProgressionForFirstPayloadWeave(
     ...progression,
     latestManifestationPath,
   };
-}
-
-function replaceExtractionSourceBlock(
-  turtle: string,
-  extractionSourcePath: string,
-  replacementBlock: string,
-): string {
-  const blocks = splitTurtleBlocks(turtle);
-  const blockIndex = findSubjectBlockIndex(blocks, extractionSourcePath);
-  if (blockIndex === -1) {
-    throw new WeaveInputError(
-      `Could not replace existing ExtractionSource block <${extractionSourcePath}>.`,
-    );
-  }
-
-  const nextBlocks = [...blocks];
-  nextBlocks[blockIndex] = replacementBlock;
-  return `${nextBlocks.join("\n\n")}\n`;
-}
-
-function renderExactExtractionSourceBlock(
-  extractionSourcePath: string,
-  sourceDesignatorPath: string,
-  sourceStatePath: string,
-  sourceEvidence: ExtractionSourceEvidenceModel | undefined,
-): string {
-  const facts: [string, string][] = [
-    ["sflo:hasTargetArtifact", `<${sourceDesignatorPath}>`],
-    ["sflo:hasRequestedTargetState", `<${sourceStatePath}>`],
-    ...toExtractionSourceEvidenceFacts(sourceEvidence),
-  ];
-
-  return `<${extractionSourcePath}> a sflo:ExtractionSource ;
-${
-    facts.map(([predicate, object], index) =>
-      `  ${predicate} ${object}${index === facts.length - 1 ? " ." : " ;"}`
-    ).join("\n")
-  }`;
-}
-
-function toExtractionSourceEvidenceFacts(
-  sourceEvidence: ExtractionSourceEvidenceModel | undefined,
-): [string, string][] {
-  if (!sourceEvidence) {
-    return [];
-  }
-
-  const facts: [string, string][] = [];
-  if (sourceEvidence.sourceStatePath !== undefined) {
-    facts.push([
-      "sflo:hasObservedSourceState",
-      `<${sourceEvidence.sourceStatePath}>`,
-    ]);
-  }
-  if (sourceEvidence.sourceManifestationPath !== undefined) {
-    facts.push([
-      "sflo:hasObservedSourceManifestation",
-      `<${sourceEvidence.sourceManifestationPath}>`,
-    ]);
-  }
-  if (sourceEvidence.sourceLocatedFilePath !== undefined) {
-    facts.push([
-      "sflo:hasObservedSourceLocatedFile",
-      `<${sourceEvidence.sourceLocatedFilePath}>`,
-    ]);
-  }
-  if (sourceEvidence.sourceLocalRelativePath !== undefined) {
-    facts.push([
-      "sflo:observedSourceLocalRelativePath",
-      `"${escapeTurtleString(sourceEvidence.sourceLocalRelativePath)}"`,
-    ]);
-  }
-  if (sourceEvidence.sourceDigest !== undefined) {
-    facts.push([
-      "sflo:observedSourceDigest",
-      `"${escapeTurtleString(sourceEvidence.sourceDigest)}"`,
-    ]);
-  }
-  if (sourceEvidence.observedAt !== undefined) {
-    facts.push([
-      "sflo:observedAt",
-      `"${escapeTurtleString(sourceEvidence.observedAt)}"`,
-    ]);
-  }
-
-  return facts;
-}
-
-function escapeTurtleString(value: string): string {
-  return value.replace(/[\b\t\n\f\r"\\]/g, (character) => {
-    switch (character) {
-      case "\b":
-        return "\\b";
-      case "\t":
-        return "\\t";
-      case "\n":
-        return "\\n";
-      case "\f":
-        return "\\f";
-      case "\r":
-        return "\\r";
-      case '"':
-        return '\\"';
-      case "\\":
-        return "\\\\";
-      default:
-        return character;
-    }
-  });
 }
 
 function toFileName(path: string): string {
